sysinv/cgts-client/cgts-client/requirements.txt
python-neutronclient
python-ceilometerclient
